feat: Challenge componet with his options without logic, only style
Adds the Card, Challenge, and QuestionBubble components to the lesson module. The Card component is responsible for rendering a card with an image, text, and status indicators. The Challenge component displays a set of options for the user to select from. The QuestionBubble component renders a question bubble with an image and text. These components are essential for building interactive lessons and quizzes.